JS-2421 Isolate SonarLint runtime API guard test
